# internal/platform/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tibi API
  description: |
    Medical consultation platform API.

    Authenticate with `POST /api/v1/auth/login`, then pass the access token as
    `Authorization: Bearer <token>`. WebSocket hub uses `?token=<jwt>` on upgrade.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /hubs/consultations:
    get:
      tags: [Realtime]
      summary: WebSocket hub
      description: |
        Upgrade to WebSocket. Auth via query param `token` (JWT).
        Clients may join consultation and queue groups after connect.
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "101":
          description: Switching Protocols
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "204":
          description: Accepted (always, to avoid enumeration)
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Confirm password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        "204":
          description: Password updated
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/auth/verify-email/confirm:
    post:
      tags: [Auth]
      summary: Confirm email verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "204":
          description: Email verified
  /api/v1/auth/verify-email/request:
    post:
      tags: [Auth]
      summary: Request a new verification email
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Sent if applicable
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout (revoke refresh token)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "204":
          description: Logged out
  /api/v1/auth/webhooks/kashier:
    post:
      tags: [Auth]
      summary: Kashier payment webhook
      description: Signature-verified webhook; not for client use.
      responses:
        "200":
          description: Accepted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
...
  /api/v1/doctors:
    get:
      tags: [Doctors]
      summary: Search doctors
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: specialty_id
          in: query
          schema:
            type: integer
            format: int64
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
  /api/v1/doctors/{id}:
    get:
      tags: [Doctors]
      summary: Get doctor by id
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: OK
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/doctors/{doctorId}/posts:
    get:
      tags: [Posts]
      summary: List posts by doctor
      parameters:
        - name: doctorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
  /api/v1/doctors/{doctorId}/free-messages:
    post:
      tags: [FreeMessages]
      summary: Submit a free message to a doctor (public)
      parameters:
        - name: doctorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sender_name, sender_email, body]
              properties:
                sender_name:
                  type: string
                sender_email:
                  type: string
                  format: email
                body:
                  type: string
      responses:
        "201":
          description: Created
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/doctors/me/free-messages:
    get:
      tags: [FreeMessages]
      summary: Doctor inbox of free messages
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/free-messages/{id}/reply:
    post:
      tags: [FreeMessages]
      summary: Reply to a free message
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body:
                  type: string
      responses:
        "200":
          description: OK
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/profile/doctor:
    patch:
      tags: [Profile]
      summary: Update doctor profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/profile/doctor/submit-for-verification:
    post:
      tags: [Profile]
      summary: Submit doctor profile for admin verification
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Submitted
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /api/v1/profile/image:
    patch:
      tags: [Profile]
      summary: Set profile image from an uploaded file
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_id]
              properties:
                file_id:
                  type: integer
                  format: int64
      responses:
        "200":
          description: OK
  /api/v1/doctor-availability/{doctorProfileId}/days:
    get:
      tags: [Schedule]
      summary: List availability days for a doctor
      parameters:
        - name: doctorProfileId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
...
  /api/v1/doctor-schedule/exceptions:
    post:
      tags: [Schedule]
      summary: Add a schedule exception
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [Patient, PendingDoctor]
        full_name:
          type: string
    AuthTokens:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        token_type:
          type: string
          example: Bearer
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthenticated:
      description: Unauthenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...